organized new components into dialog folder
