+++ website/src/components/user-management/user-list.tsx
@@ -1,5 +1,7 @@
 import React, { useState } from "react"
 import { Input } from "reakit"
+import { ConfirmationDialog } from "src/components/dialog/confirmation-dialog"
+import { EmptyDialog } from "src/components/dialog/empty-dialog"
 import {
   AllUsersQuery,
   User,
@@ -8,8 +10,6 @@ import {
   useDeleteUserMutation,
   useUpdateUserMutation,
 } from "src/graphql/dailp"
-import { ConfirmationDialog } from "../confirmation-dialog"
-import { EmptyDialog } from "../empty-dialog"
 import { RoleDropdown } from "./role-dropdown"
 import * as css from "./user-list.css"
