+++ website/src/pages/admin/manage-users.css.ts
@@ -1,5 +1,7 @@
 import { style } from "@vanilla-extract/css"
-import { space, thickness } from "src/style/constants"
+import { important } from "polished"
+import { colors, fonts, space, thickness, vspace } from "src/style/constants"
+import { std } from "src/style/utils.css"
 
 export const headerContainer = style({
   display: "flex",
@@ -19,4 +21,29 @@ export const addButton = style({
   border: thickness.none,
   cursor: "pointer",
   fontFamily: "Quattrocento Sans, sans-serif",
+})
+
+export const listHeaderContainer = style([
+  std.fullWidth,
+  important({
+    margin: 0,
+  }),
+  {
+    display: "flex",
+    position: "sticky",
+    flexFlow: "row nowrap",
+    height: vspace[1.75],
+  },
+])
+
+export const listHeaderTab = style({
+  borderRadius: 0,
+  border: "none",
+  flexGrow: 1,
+  cursor: "default",
+  fontFamily: fonts.header,
+  fontSize: "1.1rem",
+  backgroundColor: colors.secondary,
+  color: colors.secondaryContrast,
+  textAlign: "center",
 })
